requirements.txt
asgiref
boto3==1.35.26
botocore==1.35.26
certifi==2024.8.30
charset-normalizer==3.3.2
Django
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
fsspec==2024.9.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==2.1.1
pandas==2.2.3
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
tzdata
urllib3==2.2.3